+++ .gitignore
@@ -1,21 +1,2 @@
 
 weights/MealSAM.pth
-weights/sam_vit_b_01ec64.pth
-main.spec
-MealSAM_food_annotator.py
-build/main/Analysis-00.toc
-build/main/base_library.zip
-build/main/EXE-00.toc
-build/main/main.pkg
-build/main/PKG-00.toc
-build/main/PYZ-00.pyz
-build/main/PYZ-00.toc
-build/main/warn-main.txt
-build/main/xref-main.html
-build/main/localpycs/pyimod01_archive.pyc
-build/main/localpycs/pyimod02_importers.pyc
-build/main/localpycs/pyimod03_ctypes.pyc
-build/main/localpycs/pyimod04_pywin32.pyc
-build/main/localpycs/struct.pyc
-dist/main.exe
-dist/MealSAM.zip
